+++ Program.cs
@@ -1,7 +1,10 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
+using LabManager.Repositories;
 
 new DatabaseSetup();
+
+var computerRepository = new ComputerRepository();
 
 
 //Routing 
@@ -13,21 +16,10 @@ if(modelName == "Computer")
         if(modelAction == "List")
         {
                 Console.WriteLine("Computer List");
-                var connection = new SqliteConnection("Data Source=database.db");
-                connection.Open();
-
-                var command = connection.CreateCommand();
-                command.CommandText = "SELECT * FROM Computers;";
-
-                var reader = command.ExecuteReader();
-
-                while(reader.Read())
+                foreach (var computer in computerRepository.GetAll())
                 {
-                        Console.WriteLine("{0}, {1}, {2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2));
-                }
-
-                reader.Close();
-                connection.Close();  
+                        Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+                } 
         }
 
         if(modelAction == "New")
